@@ -10,7 +10,7 @@
         <RootNamespace>Octopus.Tentacle.Tests.Integration</RootNamespace>
         <RuntimeIdentifiers>win;win-x86;win-x64;linux-x64;osx-x64;linux-arm64;linux-arm;linux-musl-x64</RuntimeIdentifiers>
         <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-        <LangVersion>8</LangVersion>
+        <LangVersion>9</LangVersion>
         <Nullable>annotations</Nullable>
     </PropertyGroup>
     <PropertyGroup Condition="!$([MSBuild]::IsOSUnixLike())">
@@ -23,10 +23,9 @@
     <PropertyGroup Condition=" '$(TargetFramework)' == 'net48' ">
         <DefineConstants>$(DefineConstants);DOES_NOT_SUPPORT_CANCELLATION_ON_SOCKETS</DefineConstants>
     </PropertyGroup>
-    
-    <ItemGroup>
+
+  <ItemGroup>
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
-        <ProjectReference Include="..\Octopus.Tentacle.Client\Octopus.Tentacle.Client.csproj" />
         <ProjectReference Include="..\Octopus.Tentacle.CommonTestUtils\Octopus.Tentacle.CommonTestUtils.csproj" />
         <PackageReference Include="NSubstitute" Version="4.4.0" />
         <PackageReference Include="FluentAssertions" Version="6.7.0" />
