@@ -9,7 +9,7 @@
     <RootNamespace>Octopus.Tentacle.Tests</RootNamespace>
     <RuntimeIdentifiers>win;win-x86;win-x64;linux-x64;osx-x64;linux-arm64;linux-arm;linux-musl-x64</RuntimeIdentifiers>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <LangVersion>8</LangVersion>
+    <LangVersion>9</LangVersion>
     <Nullable>annotations</Nullable>
   </PropertyGroup>
 
