@@ -1,8 +1,11 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using System.Net.Sockets;
 using System.Threading;
 using System.Threading.Tasks;
+using Octopus.Tentacle.Tests.Integration.Support;
 using Serilog;
 
 namespace Octopus.Tentacle.Tests.Integration.Util.TcpUtils
@@ -13,21 +16,22 @@ namespace Octopus.Tentacle.Tests.Integration.Util.TcpUtils
         readonly EndPoint clientEndPoint;
         readonly Socket originSocket;
         readonly EndPoint originEndPoint;
-        readonly CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
+        readonly CancellationTokenSource cancellationTokenSource = new();
         readonly ILogger logger;
         readonly TimeSpan sendDelay;
         bool isDisposing;
         bool isDisposed;
         public bool IsPaused { get; set; }
-        
+        private Func<BiDirectionalDataTransferObserver> factory;
 
-        public TcpPump(Socket clientSocket, Socket originSocket, EndPoint originEndPoint, TimeSpan sendDelay, ILogger logger)
+        public TcpPump(Socket clientSocket, Socket originSocket, EndPoint originEndPoint, TimeSpan sendDelay, Func<BiDirectionalDataTransferObserver> factory, ILogger logger)
         {
-            this.logger = logger.ForContext<TcpPump>(); 
+            this.logger = logger.ForContext<TcpPump>();
             this.clientSocket = clientSocket ?? throw new ArgumentNullException(nameof(clientSocket));
             this.originSocket = originSocket ?? throw new ArgumentNullException(nameof(originSocket));
             this.originEndPoint = originEndPoint ?? throw new ArgumentNullException(nameof(originEndPoint));
             this.sendDelay = sendDelay;
+            this.factory = factory;
             clientEndPoint = clientSocket.RemoteEndPoint ?? throw new ArgumentException("Remote endpoint is null", nameof(clientSocket));
         }
 
@@ -44,15 +48,30 @@ namespace Octopus.Tentacle.Tests.Integration.Util.TcpUtils
 
                     try
                     {
-                        await originSocket.ConnectAsync(originEndPoint).ConfigureAwait(false);
+#if DOES_NOT_SUPPORT_CANCELLATION_ON_SOCKETS
+                        var connectAsyncCancellationTokenSource = new CancellationTokenSource();
+                        using (cancellationToken.Register(() => connectAsyncCancellationTokenSource.Cancel()))
+                        {
+                            var cancelTask = connectAsyncCancellationTokenSource.Token.AsTask();
+                            var actionTask = originSocket.ConnectAsync(originEndPoint);
 
+                            await (await Task.WhenAny(actionTask, cancelTask).ConfigureAwait(false)).ConfigureAwait(false);
+                        }
+#else
+                        await originSocket.ConnectAsync(originEndPoint, cancellationToken).ConfigureAwait(false);
+#endif
+
+                        cancellationToken.ThrowIfCancellationRequested();
+
+                        var biDirectionalCallBack = factory();
                         // If the connection was ok, then set-up a pump both ways
-                        var pump1 = Task.Run(async () => await PumpBytes(clientSocket, originSocket, new SocketPump(() => this.IsPaused, sendDelay), cancellationToken).ConfigureAwait(false), cancellationToken);
-                        var pump2 = Task.Run(async () => await PumpBytes(originSocket, clientSocket, new SocketPump(() => this.IsPaused, sendDelay), cancellationToken).ConfigureAwait(false), cancellationToken);
+                        var pump1 = Task.Run(async () => await PumpBytes(clientSocket, originSocket, new SocketPump(this, () => this.IsPaused, sendDelay, biDirectionalCallBack.DataTransferObserverClientToOrigin), cancellationToken).ConfigureAwait(false), cancellationToken);
+                        var pump2 = Task.Run(async () => await PumpBytes(originSocket, clientSocket, new SocketPump(this, () => this.IsPaused, sendDelay, biDirectionalCallBack.DataTransferObserverOriginToClient), cancellationToken).ConfigureAwait(false), cancellationToken);
 
                         // When one is finished, they are both "done" so stop them
                         await Task.WhenAny(pump1, pump2).ConfigureAwait(false);
                         logger.Verbose("Stopping connection forwarding from {ClientEndPoint} to {OriginEndPoint}.", clientEndPoint.ToString(), originEndPoint.ToString());
+
                         if (!cancellationTokenSource.IsCancellationRequested)
                         {
                             cancellationTokenSource.Cancel();
@@ -68,8 +87,8 @@ namespace Octopus.Tentacle.Tests.Integration.Util.TcpUtils
 
                     // We are done. Close everything.
                     logger.Verbose("Stopped connection forwarding from {ClientEndPoint} to {OriginEndPoint}.", clientEndPoint.ToString(), originEndPoint.ToString());
-                    clientSocket.Close();
-                    originSocket.Close();
+                    clientSocket.Close(0);
+                    originSocket.Close(0);
 
                     // Let the users know we are done
                     Stopped?.Invoke(this, EventArgs.Empty);
@@ -90,7 +109,7 @@ namespace Octopus.Tentacle.Tests.Integration.Util.TcpUtils
                 try
                 {
 
-                    var socketStatus = await socketPump.PumpBytes(readFrom, writeTo, cancellationToken);
+                    var socketStatus = await socketPump.PumpBytes(readFrom, writeTo, cancellationToken).ConfigureAwait(false);
                     if(socketStatus == SocketPump.SocketStatus.SOCKET_CLOSED) break;
                 }
                 catch (SocketException socketException)
@@ -108,6 +127,15 @@ namespace Octopus.Tentacle.Tests.Integration.Util.TcpUtils
             }
         }
 
+        public void Pause()
+        {
+            IsPaused = true;
+        }
+
+        public void UnPause()
+        {
+            IsPaused = false;
+        }
 
         public void Dispose()
         {
@@ -118,14 +146,67 @@ namespace Octopus.Tentacle.Tests.Integration.Util.TcpUtils
             cancellationTokenSource.Cancel();
             cancellationTokenSource.Dispose();
             isDisposed = true;
-            
-            clientSocket.Close();
-            originSocket.Close();
-        }
 
-        public void Pause()
-        {
-            IsPaused = true;
+            var exceptions = new List<Exception>();
+
+            try
+            {
+                clientSocket.Shutdown(SocketShutdown.Both);
+            }
+            catch (Exception ex)
+            {
+                exceptions.Add(ex);
+            }
+
+            try
+            {
+                originSocket.Shutdown(SocketShutdown.Both);
+            }
+            catch (Exception ex)
+            {
+                exceptions.Add(ex);
+            }
+
+            try
+            {
+                clientSocket.Close(0);
+            }
+            catch (Exception ex)
+            {
+                exceptions.Add(ex);
+            }
+
+            try
+            {
+                originSocket.Close(0);
+            }
+            catch (Exception ex)
+            {
+                exceptions.Add(ex);
+            }
+
+            try
+            {
+                clientSocket.Dispose();
+            }
+            catch (Exception ex)
+            {
+                exceptions.Add(ex);
+            }
+
+            try
+            {
+                originSocket.Dispose();
+            }
+            catch (Exception ex)
+            {
+                exceptions.Add(ex);
+            }
+
+            if (exceptions.Count(x => x is not ObjectDisposedException) > 0)
+            {
+                logger.Warning(new AggregateException(exceptions), "Errors occurred Disposing of the TcpPump");
+            }
         }
     }
 }
