@@ -1,3 +1,5 @@
+using System;
+using System.Collections;
 using System.Threading.Tasks;
 using FluentAssertions;
 using NUnit.Framework;
@@ -6,23 +8,48 @@ using Octopus.Tentacle.Tests.Integration.Support.Legacy;
 
 namespace Octopus.Tentacle.Tests.Integration
 {
+    public class CapabilitiesServiceInterestingTentacles : IEnumerable
+    {
+        public IEnumerator GetEnumerator()
+        {
+            return AllCombinations
+                .Of(new TentacleTypesToTest())
+                .And(
+                    TentacleVersions.Current,
+                    TentacleVersions.v5_0_4_FirstLinuxRelease,
+                    TentacleVersions.v5_0_12_AutofacServiceFactoryIsInShared,
+                    TentacleVersions.v6_3_417_LastWithScriptServiceV1Only // the autofac service is in tentacle, but tentacle does not have the capabilities service.
+                )
+                .Build();
+        }
+    }
+    
+    [IntegrationTestTimeout]
     public class CapabilitiesServiceV2Test : IntegrationTest
     {
-        [TestCase(null)] // The version of tentacle compiled from the current code.
-        [TestCase("5.0.4")] // First linux Release 9/9/2019
-        [TestCase("5.0.12")] // The autofac service was in octopus shared.
-        [TestCase("6.3.451")] // the autofac service is in tentacle, but tentacle does not have the capabilities service.
-        public async Task CapabilitiesFromAnOlderTentacleWhichHasNoCapabilitiesService_WorksWithTheBackwardsCompatabilityDecorator(string version)
+        [Test]
+        [TestCaseSource(typeof(CapabilitiesServiceInterestingTentacles))]
+        public async Task CapabilitiesFromAnOlderTentacleWhichHasNoCapabilitiesService_WorksWithTheBackwardsCompatabilityDecorator(
+            TentacleType tentacleType,
+            string? version)
         {
-            using var clientAndTentacle = await new ClientAndTentacleBuilder(TentacleType.Polling)
-                .WithTentacleVersion(version)
+            using var clientAndTentacle = await new LegacyClientAndTentacleBuilder(tentacleType)
+                .WithTentacleVersion(version!)
                 .Build(CancellationToken);
 
             var capabilities = clientAndTentacle.TentacleClient.CapabilitiesServiceV2.GetCapabilities().SupportedCapabilities;
 
             capabilities.Should().Contain("IScriptService");
             capabilities.Should().Contain("IFileTransferService");
-            capabilities.Count.Should().Be(2);
+            if (version == null)
+            {
+                capabilities.Should().Contain("IScriptServiceV2Alpha");
+                capabilities.Count.Should().Be(3);
+            }
+            else
+            {
+                capabilities.Count.Should().Be(2);
+            }
         }
     }
 }
