@@ -1,3 +1,4 @@
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Octopus.Tentacle.Tests.Integration.Support
@@ -10,7 +11,8 @@ namespace Octopus.Tentacle.Tests.Integration.Support
         /// </summary>
         /// <param name="tmp">A directory that may be used for holding the binaries or used while downloading the binaries.</param>
         /// <param name="version">The version to download.</param>
+        /// <param name="cancellationToken"></param>
         /// <returns>The full path to the tentacle executable.</returns>
-        public Task<string> GetTentacleVersion(string tmp, string version);
+        public Task<string> GetTentacleVersion(string tmp, string version, CancellationToken cancellationToken);
     }
 }
