@@ -1,29 +1,28 @@
-﻿using System;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 using Halibut;
 using Octopus.Tentacle.Contracts.Legacy;
-using Octopus.Tentacle.Tests.Integration.Util;
+using Octopus.Tentacle.Tests.Integration.Util.TcpUtils;
 
 namespace Octopus.Tentacle.Tests.Integration.Support.Legacy
 {
-    internal class ClientAndTentacleBuilder
+    internal class LegacyClientAndTentacleBuilder
     {
         private readonly TentacleType tentacleType;
         private string? tentacleVersion;
 
-        public ClientAndTentacleBuilder(TentacleType tentacleType)
+        public LegacyClientAndTentacleBuilder(TentacleType tentacleType)
         {
             this.tentacleType = tentacleType;
         }
 
-        public ClientAndTentacleBuilder WithTentacleVersion(string tentacleVersion)
+        public LegacyClientAndTentacleBuilder WithTentacleVersion(string tentacleVersion)
         {
             this.tentacleVersion = tentacleVersion;
             return this;
         }
 
-        public async Task<ClientAndTentacle> Build(CancellationToken cancellationToken)
+        public async Task<LegacyClientAndTentacle> Build(CancellationToken cancellationToken)
         {
             // Server
             var serverHalibutRuntime = new HalibutRuntimeBuilder()
@@ -37,34 +36,41 @@ namespace Octopus.Tentacle.Tests.Integration.Support.Legacy
             var server = new Server(serverHalibutRuntime, serverListeningPort);
 
             // Port Forwarder
-            var portForwarder = PortForwarderBuilder.ForwardingToLocalPort(serverListeningPort).Build();
+            PortForwarder portForwarder;
             RunningTentacle runningTentacle;
+            ServiceEndPoint tentacleEndPoint;
 
             // Tentacle
             var temporaryDirectory = new TemporaryDirectory();
             var tentacleExe = string.IsNullOrWhiteSpace(tentacleVersion) ?
                 TentacleExeFinder.FindTentacleExe() :
-                await TentacleFetcher.GetTentacleVersion(temporaryDirectory.DirectoryPath, tentacleVersion);
+                await TentacleFetcher.GetTentacleVersion(temporaryDirectory.DirectoryPath, tentacleVersion, cancellationToken);
 
             if (tentacleType == TentacleType.Polling)
             {
+                portForwarder = PortForwarderBuilder.ForwardingToLocalPort(serverListeningPort).Build();
+
                 runningTentacle = await new PollingTentacleBuilder(portForwarder.ListeningPort, Certificates.ServerPublicThumbprint)
                     .WithTentacleExe(tentacleExe)
                     .Build(cancellationToken);
+
+                tentacleEndPoint = new ServiceEndPoint(runningTentacle.ServiceUri, runningTentacle.Thumbprint);
             }
             else
             {
                 runningTentacle = await new ListeningTentacleBuilder(Certificates.ServerPublicThumbprint)
                     .WithTentacleExe(tentacleExe)
                     .Build(cancellationToken);
+
+                portForwarder = new PortForwarderBuilder(runningTentacle.ServiceUri).Build();
+
+                tentacleEndPoint = new ServiceEndPoint(portForwarder.PublicEndpoint, runningTentacle.Thumbprint);
             }
 
-            var tentacleEndPoint = new ServiceEndPoint(runningTentacle.ServiceUri, runningTentacle.Thumbprint);
-
-            var tentacleClient = new TentacleClientBuilder(server.ServerHalibutRuntime, tentacleEndPoint)
+            var tentacleClient = new LegacyTentacleClientBuilder(server.ServerHalibutRuntime, tentacleEndPoint)
                 .Build(cancellationToken);
 
-            return new ClientAndTentacle(server, portForwarder, runningTentacle, tentacleClient, temporaryDirectory);
+            return new LegacyClientAndTentacle(server, portForwarder, runningTentacle, tentacleClient, temporaryDirectory);
         }
     }
 }
