@@ -4,9 +4,9 @@ using Octopus.Tentacle.Contracts.Capabilities;
 
 namespace Octopus.Tentacle.Tests.Integration.Support.Legacy
 {
-    public class TentacleClient
+    public class LegacyTentacleClient
     {
-        public TentacleClient(IScriptService scriptService, IFileTransferService fileTransferService, ICapabilitiesServiceV2 capabilitiesServiceV2)
+        public LegacyTentacleClient(IScriptService scriptService, IFileTransferService fileTransferService, ICapabilitiesServiceV2 capabilitiesServiceV2)
         {
             ScriptService = scriptService;
             FileTransferService = fileTransferService;
@@ -15,7 +15,6 @@ namespace Octopus.Tentacle.Tests.Integration.Support.Legacy
 
         public IScriptService ScriptService { get; }
         public IFileTransferService FileTransferService { get; }
-
         public ICapabilitiesServiceV2 CapabilitiesServiceV2 { get; }
     }
 }
