@@ -25,26 +25,41 @@ namespace Octopus.Tentacle.Tests.Integration.Support
             AddCertificateToTentacle(tentacleExe, instanceName, CertificatePfxPath, tempDirectory, cancellationToken);
             ConfigureTentacleToListen(configFilePath);
 
-            return await StartTentacle(
+            var runningTentacle = await StartTentacle(
                 null,
                 tentacleExe,
                 instanceName,
                 tempDirectory,
                 TentacleThumbprint,
                 cancellationToken);
+
+            SetThePort(configFilePath, runningTentacle.ServiceUri.Port);
+
+            return runningTentacle;
         }
 
         private void ConfigureTentacleToListen(string configFilePath)
         {
-            var writableTentacleConfiguration = GetWritableTentacleConfiguration(configFilePath);
-
-            writableTentacleConfiguration.AddOrUpdateTrustedOctopusServer(new OctopusServerConfiguration(ServerThumbprint)
+            WithWritableTentacleConfiguration(configFilePath, writableTentacleConfiguration =>
             {
-                CommunicationStyle = CommunicationStyle.TentaclePassive,
-            });
+                writableTentacleConfiguration.AddOrUpdateTrustedOctopusServer(new OctopusServerConfiguration(ServerThumbprint)
+                {
+                    CommunicationStyle = CommunicationStyle.TentaclePassive,
+                });
 
-            writableTentacleConfiguration.SetServicesPortNumber(0); // Find a random available port
-            writableTentacleConfiguration.SetNoListen(false);
+                writableTentacleConfiguration.SetApplicationDirectory(Path.Combine(new DirectoryInfo(configFilePath).Parent.FullName, "appdir"));
+
+                writableTentacleConfiguration.SetServicesPortNumber(0); // Find a random available port
+                writableTentacleConfiguration.SetNoListen(false);
+            });
+        }
+
+        private void SetThePort(string configFilePath, int port)
+        {
+            WithWritableTentacleConfiguration(configFilePath, writableTentacleConfiguration =>
+            {
+                writableTentacleConfiguration.SetServicesPortNumber(port);
+            });
         }
     }
 }
