@@ -7,13 +7,13 @@ namespace Octopus.Tentacle.Tests.Integration.Support
     {
         public IEnumerator GetEnumerator()
         {
-            yield return new object[] { TentacleType.Polling, null };
-            yield return new object[] { TentacleType.Polling, "5.0.15" };
-            yield return new object[] { TentacleType.Polling, "6.3.417" };
+            yield return new object[] { TentacleType.Polling, TentacleVersions.Current };
+            yield return new object[] { TentacleType.Polling, TentacleVersions.v5_0_15_LastOfVersion5 };
+            yield return new object[] { TentacleType.Polling, TentacleVersions.v6_3_417_LastWithScriptServiceV1Only };
 
-            yield return new object[] { TentacleType.Listening, null };
-            yield return new object[] { TentacleType.Listening, "5.0.15" };
-            yield return new object[] { TentacleType.Listening, "6.3.417" };
+            yield return new object[] { TentacleType.Listening, TentacleVersions.Current };
+            yield return new object[] { TentacleType.Listening, TentacleVersions.v5_0_15_LastOfVersion5 };
+            yield return new object[] { TentacleType.Listening, TentacleVersions.v6_3_417_LastWithScriptServiceV1Only };
         }
     }
 }
