@@ -15,7 +15,8 @@ namespace Octopus.Tentacle.Tests.Capabilities
 
             capabilities.Should().Contain("IScriptService");
             capabilities.Should().Contain("IFileTransferService");
-            capabilities.Count.Should().Be(2);
+            capabilities.Should().Contain("IScriptServiceV2Alpha");
+            capabilities.Count.Should().Be(3);
         }
     }
 }
