@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using Octopus.Tentacle.Contracts;
 using Octopus.Tentacle.Contracts.Capabilities;
+using Octopus.Tentacle.Contracts.ScriptServiceV2;
 
 namespace Octopus.Tentacle.Services.Capabilities
 {
@@ -9,7 +10,7 @@ namespace Octopus.Tentacle.Services.Capabilities
     {
         public CapabilitiesResponseV2 GetCapabilities()
         {
-            return new CapabilitiesResponseV2(new List<string>() {nameof(IScriptService), nameof(IFileTransferService)});
+            return new CapabilitiesResponseV2(new List<string>() {nameof(IScriptService), nameof(IFileTransferService), nameof(IScriptServiceV2) + "Alpha"});
         }
     }
 }
